# Servers/swagger.yaml
openapi: 3.0.0
info:
  title: VerifyWise API
  description: AI Governance Platform API
  version: 2.0.0
servers:
  - url: /api
    description: Main API server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
...
  /autoDrivers:
    post:
      tags:
        - Demo Data
      summary: Post Auto Driver
      operationId: postAutoDriver
      security:
        - bearerAuth: []
      description: "Requires role: Admin"
      requestBody:
        content:
          application/json:
            schema:
              type: object
      responses:
        "201":
          description: Created successfully
        "401":
          description: Unauthorized
        "403":
          description: Forbidden - insufficient role
        "500":
          description: Internal server error
    delete:
      tags:
        - Demo Data
      summary: Delete Auto Driver
      operationId: deleteAutoDriver
      security:
        - bearerAuth: []
      description: "Requires role: Admin"
      responses:
        "200":
          description: Deleted successfully
        "401":
          description: Unauthorized
        "403":
          description: Forbidden - insufficient role
        "500":
          description: Internal server error
...